+++ src/app/app.routes.ts
@@ -1,26 +1,27 @@
 import { Routes } from '@angular/router';
-import { unauthGuard } from './guards/unauth.guard';
+import { authGuard } from './core/guards/auth.guard';
+import { unauthGuard } from './core/guards/unauth.guard';
 
 export const routes: Routes = [
   {
     path: 'login',
-    title: 'Login',
+    title: 'Login | financee.me',
     canActivate: [unauthGuard],
     loadComponent: () =>
-      import('./pages/login/login.component').then((m) => m.LoginComponent),
+      import('./features/login/login.component').then((m) => m.LoginComponent),
   },
   {
     path: 'register',
-    title: 'Login',
+    title: 'Registrar | financee.me',
     loadComponent: () =>
-      import('./pages/register/register.component').then((m) => m.RegisterComponent),
+      import('./features/register/register.component').then((m) => m.RegisterComponent),
   },
   {
     path: '',
     title: 'financee.me',
-    // canActivate: [authGuard],
+    canActivate: [authGuard],
     loadComponent: () =>
-      import('./components/navbar/navbar.component').then(
+      import('./core/components/navbar/navbar.component').then(
         (m) => m.NavbarComponent
       ),
     children: [
@@ -28,19 +29,19 @@ export const routes: Routes = [
         path: 'overview',
         title: 'financee.me',
         loadComponent: () =>
-          import('./pages/overview/overview.component').then((m) => m.OverviewComponent),
+          import('./features/overview/overview.component').then((m) => m.OverviewComponent),
       },
       {
         path: 'finance',
         title: 'Transações | financee.me',
         loadComponent: () =>
-          import('./pages/finance/finance.component').then((m) => m.FinanceComponent),
+          import('./features/finance/finance.component').then((m) => m.FinanceComponent),
       },
       {
         path: 'plans',
         title: 'Planos | financee.me',
         loadComponent: () =>
-          import('./pages/plans/plans.component').then((m) => m.PlansComponent),
+          import('./features/plans/plans.component').then((m) => m.PlansComponent),
       },
       {
         path: '**',