+++ src/environments/environment.production.ts
@@ -1,5 +1,5 @@
 export const environment = {
   production: true,
   base_url: 'https://finance-api.iamgui.dev/',
-  version: '1.2.1',
+  version: '1.3.0',
 };